microbench-dashboard: initial commit and clean-up
This repo is a subset of: https://github.com/golang/build

The intention is to reuse the performance dashboards to display microbenchmark
results stored in InfluxDB.

All LICENSES and PATENTS have been kept intact. Refer to the original repository
for more information.

Everything not related to the performance dashboards will be removed from this
repo. Some areas of the code will be repurposed to match custom infrastructure and
tooling requirements.