perf: remove git history
Instead of grabbing a history from Git, the commits will be inferred from the
benchmarks. Pulling this history was useful for displaying missing commits, but
since the microbenchmark suite does not run regularly in this case it's better
to display only commits than a microbenchmark suite has been run for.